+++ MarketOrderProvider.h
@@ -37,8 +37,8 @@ namespace Evernus
         virtual OrderList getBuyOrders(Character::IdType characterId) const = 0;
         virtual OrderList getArchivedOrders(Character::IdType characterId, const QDateTime &from, const QDateTime &to) const = 0;
 
-        virtual OrderList getSellOrdersForCorporation(uint corporationId) const = 0;
-        virtual OrderList getBuyOrdersForCorporation(uint corporationId) const = 0;
-        virtual OrderList getArchivedOrdersForCorporation(uint corporationId, const QDateTime &from, const QDateTime &to) const = 0;
+        virtual OrderList getSellOrdersForCorporation(quint64 corporationId) const = 0;
+        virtual OrderList getBuyOrdersForCorporation(quint64 corporationId) const = 0;
+        virtual OrderList getArchivedOrdersForCorporation(quint64 corporationId, const QDateTime &from, const QDateTime &to) const = 0;
     };
 }
