+++ WalletTransaction.h
@@ -72,8 +72,8 @@ namespace Evernus
         WalletJournalEntry::IdType getJournalId() const;
         void setJournalId(WalletJournalEntry::IdType id);
 
-        uint getCorporationId() const noexcept;
-        void setCorporationId(uint id) noexcept;
+        quint64 getCorporationId() const noexcept;
+        void setCorporationId(quint64 id) noexcept;
 
         bool isIgnored() const noexcept;
         void setIgnored(bool flag) noexcept;
@@ -92,7 +92,7 @@ namespace Evernus
         quint64 mLocationId = 0;
         Type mType = Type::Buy;
         WalletJournalEntry::IdType mJournalId = WalletJournalEntry::invalidId;
-        uint mCorporationId = 0;
+        quint64 mCorporationId = 0;
         bool mIgnored = false;
     };
